requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
paapi5-python-sdk==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.0a3
